add the qualifier endpoint and fix matches
